% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByPlace.R
\name{extractPlace}
\alias{extractPlace}
\title{Extract a 10s place}
\usage{
extractPlace(mHalves, place, call = rlang::caller_env())
}
\arguments{
\item{mHalves}{\code{matrix} A 2-column character matrix of numbers split on the
decimal point.}

\item{place}{The 10s place to return.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
An integer vector of digits from the numbers, at the given place.
}
\description{
Extract a 10s place
}
\keyword{internal}
