% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.digitpref-package.R
\docType{package}
\name{gsm.digitpref-package}
\alias{gsm.digitpref}
\alias{gsm.digitpref-package}
\title{gsm.digitpref: Identify and Visualize Digit Preference}
\description{
Visualization the distributions of digits in data, optionally grouped by site or other grouping variable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openrbqm.github.io/gsm.digitpref/}
  \item \url{https://github.com/OpenRBQM/gsm.digitpref}
  \item Report bugs at \url{https://github.com/OpenRBQM/gsm.digitpref/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jon.harmon@atorusresearch.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jeremy Wildfire \email{jwildfire@gmail.com}
}

Other contributors:
\itemize{
  \item Shuo Wang \email{shuo.wang@gilead.com} [contributor]
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
