% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByPlace.R
\name{numToChrHalves}
\alias{numToChrHalves}
\title{Divide a vector at a decimal point}
\usage{
numToChrHalves(x, x_arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{\code{numeric} A vector of values to split.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A 2-column character matrix, where the first column contains
digits before the decimal point, and the second column contains digits
after the decimal point (or \code{NA} if there weren't any decimal points).
}
\description{
Divide a vector at a decimal point
}
\keyword{internal}
