% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDigitCounts.R
\name{plotDigitCounts}
\alias{plotDigitCounts}
\title{Plot digit distribution by group}
\usage{
plotDigitCounts(
  dfDigits,
  colGroup,
  colPlace,
  strGroupLabel = rlang::as_name({
     {
         colGroup
     }
 }),
  strPlaceLabel = rlang::as_name({
     {
         colPlace
     }
 }),
  strChartTitle =
    glue::glue("Digit Distribution for {strPlaceLabel} of Data by {strGroupLabel}"),
  scaleDigitPalette = ggplot2::scale_fill_brewer(palette = "Set3"),
  themePlot = ggplot2::theme_bw()
)
}
\arguments{
\item{dfDigits}{\code{data.frame} A data.frame with a grouping column and at least
one column for a digit place.}

\item{colGroup}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> A column to
group by.}

\item{colPlace}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> A column with
integer values between 0 and 9.}

\item{strGroupLabel}{\code{character} An optional name to use for the grouping
column in labels.}

\item{strPlaceLabel}{\code{character} An optional name to use for the place column
in labels.}

\item{strChartTitle}{\code{character} An optional title for the chart.}

\item{scaleDigitPalette}{An optional \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_fill_discrete()}} palette
to use for the digits.}

\item{themePlot}{An optional \code{\link[ggplot2:theme]{ggplot2::theme()}} to apply to the plot.}
}
\description{
Create a stacked bar chart of digit distribution, with a grouping variable on
the x axis, and the distribution of digits in a particular digit place on the
y axis.
}
\examples{
sample_data <- data.frame(
  siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
  onesPlace = sample(0:9, 1000, replace = TRUE)
)
plotDigitCounts(sample_data, siteID, onesPlace)
plotDigitCounts(sample_data, siteID, onesPlace, "Site", "One Place")
plotDigitCounts(sample_data, siteID, onesPlace, scaleDigitPalette = NULL)
}
