% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlaceDigits.R
\name{plotPlaceDigits}
\alias{plotPlaceDigits}
\title{Plot Digit Distribution for a Place by Group}
\usage{
plotPlaceDigits(
  dfData,
  colData,
  colGroup,
  power_10,
  strGroupLabel = colGroup,
  strPlaceLabel = powerToOrdinal(power_10, strAfter = "Place"),
  strChartTitle =
    glue::glue("Digit Distribution for {strPlaceLabel} of Data by {strGroupLabel}"),
  scaleDigitPalette = ggplot2::scale_fill_brewer(palette = "Set3"),
  themePlot = ggplot2::theme_bw()
)
}
\arguments{
\item{dfData}{\code{data.frame} A data.frame with a grouping column and at least
one numeric data column.}

\item{colData}{\code{character} The name of a column that contains numeric data.}

\item{colGroup}{\code{character} The name of a column to group by.}

\item{power_10}{\code{integer} The 10s place to check, as a power of 10.}

\item{strGroupLabel}{\code{character} An optional name to use for the grouping
column in labels.}

\item{strPlaceLabel}{\code{character} An optional name to use for the place column
in labels.}

\item{strChartTitle}{\code{character} An optional title for the chart.}

\item{scaleDigitPalette}{An optional \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_fill_discrete()}} palette
to use for the digits.}

\item{themePlot}{An optional \code{\link[ggplot2:theme]{ggplot2::theme()}} to apply to the plot.}
}
\description{
Create a stacked bar chart of digit distribution, with a grouping variable on
the x axis, and the distribution of digits in a particular digit place on the
y axis.
}
\examples{
sample_data <- tibble::tibble(
  siteID = sample(c("siteA", "siteB", "siteC"), 1000, replace = TRUE),
  labResults = runif(1000)*10
)
plotPlaceDigits(sample_data, "labResults", "siteID", -2)
}
