% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerToOrdinal.R
\name{powerToOrdinal}
\alias{powerToOrdinal}
\title{Convert a Power of Ten to Words}
\usage{
powerToOrdinal(
  power_10,
  strCase = c("Title", "lower", "UPPER"),
  strAfter = character()
)
}
\arguments{
\item{power_10}{\code{integer} The 10s place to check, as a power of 10.}

\item{strCase}{\code{character} Whether to display text in "Title" case, "lower"
case, or "UPPER" case.}

\item{strAfter}{\code{character} Additional text to add after the ordinal, such as
"place" or "digit".}
}
\value{
The tens place expressed in words.
}
\description{
This convenience function converts a power of ten to words, such as "Ones
Place" or "hundredths digit".
}
\examples{
powerToOrdinal(0)
powerToOrdinal(3)
powerToOrdinal(-2)
powerToOrdinal(-2, strAfter = "digit")
}
