% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByPlace.R
\name{splitByPlace}
\alias{splitByPlace}
\title{Split a vector into a table of place values}
\usage{
splitByPlace(
  x,
  max_power_10 = 2,
  min_power_10 = -2,
  x_arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{\code{numeric} A vector of values to evaluate.}

\item{max_power_10}{\code{integer} The highest 10s place to check, as a power of
10.}

\item{min_power_10}{\code{integer} The lowest 10s place to check, as a power of
10. Must be lower than \code{max_power_10}.}

\item{x_arg}{\code{character} The name of \code{x} to use in the output table and error
messages. By default this is automatically determined from the calling
context, and you can likely leave it as-is.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with a column for the original vector (named
using \code{x_arg}) and columns for each power of 10 between \code{max_power_10} and
\code{min_power_10}.
}
\description{
Generate a table of places (by power of 10, with column names such as "10^2"
and "10^-2") from a vector of numeric values.
}
\examples{
random_data <- runif(10, 0, 1000)
splitByPlace(random_data)
splitByPlace(random_data, 2, -5)
splitByPlace(random_data, 2, -5, "measurements")
}
