% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByPlace.R
\name{makePlacesList}
\alias{makePlacesList}
\title{Split a vector into a named list of place values}
\usage{
makePlacesList(
  x,
  intPlaces,
  x_arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{\code{numeric} A vector of values to evaluate.}

\item{intPlaces}{\code{integer} A vector of 10's places to include, from highest
10's place to lowest.}

\item{x_arg}{\code{character} The name of \code{x} to use in the output table and error
messages. By default this is automatically deteermined from the calling
context, and you can likely leave it as-is.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A named list of vectors of digits. Each element of the list contains
a value for each value of \code{x}, at that 10's place (given in the name of
that element).
}
\description{
This is the workhorse behind \code{\link[=splitByPlaces]{splitByPlaces()}}.
}
\keyword{internal}
