% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared.R
\name{shared-params}
\alias{shared-params}
\title{Parameters used in multiple functions}
\arguments{
\item{colData}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> A data that
contains numeric data.}

\item{colGroup}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> A column to
group by.}

\item{colPlace}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}> A column with
integer values between 0 and 9.}

\item{dfData}{\code{data.frame} A data.frame with a grouping column and at least
one numeric data column.}

\item{dfDigits}{\code{data.frame} A data.frame with a grouping column and at least
one column for a digit place.}

\item{max_power_10}{\code{integer} The highest 10s place to check, as a power of
10.}

\item{min_power_10}{\code{integer} The lowest 10s place to check, as a power of
10. Must be lower than \code{max_power_10}.}

\item{power_10}{\code{integer} The 10s place to check, as a power of 10.}

\item{scaleDigitPalette}{An optional \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_fill_discrete()}} palette
to use for the digits.}

\item{strGroupLabel}{\code{character} An optional name to use for the grouping
column in labels.}

\item{strPlaceLabel}{\code{character} An optional name to use for the place column
in labels.}

\item{strChartTitle}{\code{character} An optional title for the chart.}

\item{strCase}{\code{character} Whether to display text in "Title" case, "lower"
case, or "UPPER" case.}

\item{themePlot}{An optional \code{\link[ggplot2:theme]{ggplot2::theme()}} to apply to the plot.}

\item{x}{\code{numeric} A vector of values to evaluate.}

\item{x_arg}{\code{character} The name of \code{x} to use in the output table and error
messages. By default this is automatically determined from the calling
context, and you can likely leave it as-is.}
}
\description{
Reused parameter definitions are gathered here for easier usage.
}
\keyword{internal}
