% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByPlace.R
\name{showDigits}
\alias{showDigits}
\title{Set up display of digits for a column in a tibble}
\usage{
showDigits(x, min_power_10)
}
\arguments{
\item{x}{\code{numeric} The vector to display.}

\item{min_power_10}{\code{integer} The minimum power of 10 to focus on. We show
one place past this value, to avoid confusion from rounding.}
}
\value{
A special numeric vector class from \code{\link[tibble:num]{tibble::num()}} to display the number
}
\description{
Set up display of digits for a column in a tibble
}
\keyword{internal}
